% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLossSource.R
\docType{data}
\name{DataLossSource}
\alias{DataLossSource}
\title{Loss sources data}
\format{
A data frame with four sources of loss, one in each column.
}
\usage{
data(DataLossSource)
}
\description{
Example with data from loss sources .
}
\references{
DEMOLIN-LEITE, G.L., 2021. Importance indice: loss estimates and solution effectiveness on
production. Cuban Journal of Agricultural Science, vol. 55, no. 2, pp. 1-7.
<http://scielo.sld.cu/pdf/cjas/v55n2/2079-3480-cjas-55-02-e10.pdf>

DEMOLIN-LEITE, G.L., 2024. Do arthropods and diseases affect the production of fruits on
Caryocar brasiliense Camb. (Malpighiales: Caryocaraceae)?
Brazilian Journal of Biology, vol. 84, pp. e253215. <https://doi.org/10.1590/1519-6984.253215>
}
\author{
Germano Leao Demolin Leite : \email{germano.demolin@gmail.com}

Alcinei Mistico Azevedo : \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
