% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_LossSource.R
\name{Distribution_LossSource}
\alias{Distribution_LossSource}
\title{Loss source distribution information}
\usage{
Distribution_LossSource(DataLoss)
}
\arguments{
\item{DataLoss}{It is an matrix object containing data from loss sources.}
}
\value{
Return distribution of sources of loss: aggregate, random or regular.
}
\description{
Indicates the distribution of sources of loss: aggregate, random or regular.
}
\examples{
library(ImportanceIndice)
data("DataLossSource")
data("DataSolutionSource")

Distribution_LossSource(DataLossSource)
Distribution_SolutionSource(DataSolutionSource)

}
\seealso{
\code{\link{EffectivenessOfSolution}} ,  \code{\link{NonAttentionLevel}}  ,  \code{\link{LossSource}}
}
\author{
Germano Leao Demolin-Leite (Instituto de Ciencias Agrarias da UFMG) \cr
Alcinei Mistico Azevedo (Instituto de Ciencias Agrarias da UFMG)
}
