% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_SolutionSource.R
\name{Distribution_SolutionSource}
\alias{Distribution_SolutionSource}
\title{Solution source distribution information}
\usage{
Distribution_SolutionSource(SolutionData)
}
\arguments{
\item{SolutionData}{It is an matrix object containing data from solution sources.}
}
\value{
Return distribution of sources of solution: aggregate, random or regular.
}
\description{
Indicates the distribution of sources of solution: aggregate, random or regular.
}
\examples{
library(ImportanceIndice)
data("DataLossSource")
data("DataSolutionSource")

Distribution_LossSource(DataLossSource)
Distribution_SolutionSource(DataSolutionSource)
}
\seealso{
\code{\link{EffectivenessOfSolution}} ,  \code{\link{NonAttentionLevel}}  ,  \code{\link{LossSource}}
}
\author{
Germano Leao Demolin-Leite (Instituto de Ciencias Agrarias da UFMG) \cr
Alcinei Mistico Azevedo (Instituto de Ciencias Agrarias da UFMG)
}
